+++ src/components/index.js
@@ -256,6 +256,9 @@ export {
   EuiPageHeaderContent,
   EuiPageHeaderSection,
   EuiPageSideBar,
+  EuiPageSideBarProps,
+  EuiPageTemplate,
+  EuiPageTemplateProps,
 } from './page';
 
 export { EuiPagination, EuiPaginationButton } from './pagination';