+++ CHANGELOG.md
@@ -8,6 +8,23 @@
 - Fixed the return type of `getDefaultEuiMarkdownUiPlugins` ([#4567](https://github.com/elastic/eui/pull/4567))
 - Fixed inverse handling of boolean sorting in `EuiDataGrid` ([#4561](https://github.com/elastic/eui/pull/4561))
 
+## Feature: EuiPageTemplate ([#4517](https://github.com/elastic/eui/pull/4517))
+
+- Added new `EuiPageTemplate` component as a shortcut for creating the different types of page layout patterns
+- Added props `grow` and `direction` to `EuiPage`
+- Added props `panelled`, `panelProps`, and `paddingSize` to `EuiPageBody`
+- Added props `restrictWidth` and `paddingSize` to `EuiPageBody`
+- Added prop `paddingSize` to `EuiPageHeader`
+- Updated `tabs` prop of `EuiPageHeaderContent` to render `large` size
+- Added prop `sticky` to `EuiPageSideBar`
+- Added Sass variable `$euiPageSidebarMinWidth` for changing default `min-width` of `EuiPageSideBar`
+- Added `href` and other anchor props to `EuiHeaderSectionItemButton`
+
+**Bug fixes**
+
+- Fixed horizontal overflow of `EuiPageContent`
+- Fixed horizontal overflow of `EuiBreadcrumbs`
+
 ## [`31.8.0`](https://github.com/elastic/eui/tree/v31.8.0)
 
 - Reverted part of [#4509](https://github.com/elastic/eui/pull/4509) and returned `EuiDataGrid`'s background content area to an empty shade ([#4542](https://github.com/elastic/eui/pull/4542))